Bugfix: Fixed a segmentation fault in confirm_ui()
confirm_ui() free()ed its msg. This resulted in attemps to free const
strings or and to invalid memoory accesses (because some other functions
wanted to reuse the parameter which they gave to confirm_ui() after
calling this function).

Resolved by removing the call to free() from confirm_ui() and adding it
after the calls to confirm_ui(9, if necessary.